+++ src/main/java/com/example/surveyapp/domain/surveyanswer/infrastructure/SurveyOptionsAnswerJpaRepository.java
@@ -1,6 +1,6 @@
 package com.example.surveyapp.domain.surveyanswer.infrastructure;
 
-import com.example.surveyapp.domain.surveyanswer.domain.model.SurveyOptionsAnswer;
+import com.example.surveyapp.domain.surveyanswer.domain.model.entity.SurveyOptionsAnswer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface SurveyOptionsAnswerJpaRepository extends JpaRepository<SurveyOptionsAnswer, Long> {
