+++ src/main/java/com/example/surveyapp/domain/surveyanswer/infrastructure/SurveyTextAnswerJpaRepository.java
@@ -1,6 +1,6 @@
 package com.example.surveyapp.domain.surveyanswer.infrastructure;
 
-import com.example.surveyapp.domain.surveyanswer.domain.model.SurveyTextAnswer;
+import com.example.surveyapp.domain.surveyanswer.domain.model.entity.SurveyTextAnswer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface SurveyTextAnswerJpaRepository extends JpaRepository<SurveyTextAnswer, Long> {
