+++ src/main/java/com/example/surveyapp/domain/surveyanswer/infrastructure/SurveyTextAnswerRepositoryImpl.java
@@ -1,6 +1,6 @@
 package com.example.surveyapp.domain.surveyanswer.infrastructure;
 
-import com.example.surveyapp.domain.surveyanswer.domain.model.SurveyTextAnswer;
+import com.example.surveyapp.domain.surveyanswer.domain.model.entity.SurveyTextAnswer;
 import com.example.surveyapp.domain.surveyanswer.domain.repository.SurveyTextAnswerRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
